feat: readline done and done all functions

+++ src/commands/rm.js
@@ -1,6 +1,7 @@
 import { unlink } from 'fs/promises'
 import { getFileName, parseArgs } from "../utils.js";
 import { resolve } from 'path'
+import {app} from "../app.js";
 // import { messenger, parseSpace } from "../utils/utils.js";
 
 
@@ -8,12 +9,10 @@ import { resolve } from 'path'
 export const rm = async ( args ) => {
     let [from, to] = parseArgs( args )
 
+    if ( from === '' ) return app.printMessage('inval')
+
     const filePath = resolve( from )
-    try {
-        await unlink( filePath )
-    } catch (e) {
-        // messenger('fail')
-    }
+    await unlink( filePath )
 }
 
 // rm path_to_file
